further work on splitting requestData into two classes
